test: move allocation test ATTRIBUTES to conftest.py as fixture
